backend_v2/requirements.txt
aiofiles==0.6.0
certifi==2021.5.30
chardet==4.0.0
demjson==2.2.4
fastapi==0.65.1
fastapi-offline==1.1.0
idna==2.10
pydantic==1.8.2
PyExecJS==1.5.1
pymongo==3.11.4
python-multipart==0.0.5
pytz==2021.1
requests==2.25.1
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.6
